+++ test/HappyPheanstalkDriverTest.php
@@ -12,6 +12,7 @@
 
 namespace PMG\Queue\Driver;
 
+use Pheanstalk\Contract\PheanstalkInterface;
 use Pheanstalk\Exception\ServerException;
 use PMG\Queue\SimpleMessage;
 use PMG\Queue\DefaultEnvelope;
@@ -25,14 +26,23 @@ use PMG\Queue\Driver\Pheanstalk\PheanstalkError;
  */
 class HappyPheanstalkDriverTest extends PheanstalkTestCase
 {
-    private $conn, $serializer, $driver, $seenTubes = [];
+    private PheanstalkInterface $conn;
 
-    public function testDequeueReturnsNullWhenNoJobsAreFound()
+    private NativeSerializer $serializer;
+
+    private PheanstalkDriver $driver;
+
+    /**
+     * @var string[]
+     */
+    private array $seenTubes = [];
+
+    public function testDequeueReturnsNullWhenNoJobsAreFound() : void
     {
         $this->assertNull($this->driver->dequeue($this->randomTube()));
     }
 
-    public function testJobsCanBeEnqueuedAndDequeuedAndRemovedWithAck()
+    public function testJobsCanBeEnqueuedAndDequeuedAndRemovedWithAck() : void
     {
         $this->expectException(ServerException::class);
         $this->expectExceptionMessage('NOT_FOUND');
@@ -53,7 +63,7 @@ class HappyPheanstalkDriverTest extends PheanstalkTestCase
         $this->conn->statsJob($env2->getJob());
     }
 
-    public function testJobsCanBeEnqueuedDequeuedAndRetriedWithRetry()
+    public function testJobsCanBeEnqueuedDequeuedAndRetriedWithRetry() : void
     {
         $tube = $this->randomTube();
 
@@ -74,7 +84,7 @@ class HappyPheanstalkDriverTest extends PheanstalkTestCase
     /**
      * @group regression
      */
-    public function testRetriedJobsDoNoPutSerializePheanstalkEnvelopes()
+    public function testRetriedJobsDoNoPutSerializePheanstalkEnvelopes() : void
     {
         $tube = $this->randomTube();
 
@@ -91,7 +101,7 @@ class HappyPheanstalkDriverTest extends PheanstalkTestCase
         $this->assertInstanceOf(DefaultEnvelope::class, $env3->getWrappedEnvelope());
     }
 
-    public function testJobsAreBuriedWithRetry()
+    public function testJobsAreBuriedWithRetry() : void
     {
         $tube = $this->randomTube();
 
@@ -110,7 +120,7 @@ class HappyPheanstalkDriverTest extends PheanstalkTestCase
         $this->assertEquals('buried', $res['state']);
     }
 
-    public function testFailWithADeleteFailureStrategyRemovesTheJob()
+    public function testFailWithADeleteFailureStrategyRemovesTheJob() : void
     {
         $driver = new PheanstalkDriver($this->conn, $this->serializer, new ArrayOptions([
             ArrayOptions::RESERVE_TIMEOUT => 1,
@@ -132,7 +142,7 @@ class HappyPheanstalkDriverTest extends PheanstalkTestCase
         $this->conn->statsJob($env2->getJob());
     }
 
-    public function testJobsCanBeReleasedAfterBeingReserved()
+    public function testJobsCanBeReleasedAfterBeingReserved() : void
     {
         $tube = $this->randomTube();