+++ test/Pheanstalk/ArrayOptionsTest.php
@@ -16,30 +16,30 @@ use PMG\Queue\Driver\PheanstalkTestCase;
 
 class ArrayOptionsTest extends PheanstalkTestCase
 {
-    private $options;
+    private ArrayOptions $options;
 
-    public function testGetMessageOptionErrorsWhenGivenAnInvalidOptionName()
+    public function testGetMessageOptionErrorsWhenGivenAnInvalidOptionName() : void
     {
         $this->expectException(MissingOption::class);
 
         $this->options->getMessageOption(__METHOD__, new class() {});
     }
 
-    public function testGetMessageOptionReturnsOptionValueWhenOptionExists()
+    public function testGetMessageOptionReturnsOptionValueWhenOptionExists() : void
     {
         $option = $this->options->getMessageOption('test', new class() {});
 
         $this->assertSame(123, $option);
     }
 
-    public function testGetGlobalOptionErrorsWhenGivenAnInvalidOptionName()
+    public function testGetGlobalOptionErrorsWhenGivenAnInvalidOptionName() : void
     {
         $this->expectException(MissingOption::class);
 
         $this->options->getGlobalOption(__METHOD__);
     }
 
-    public function testGetGlobalOptionReturnsOptionValueWhenOptionExists()
+    public function testGetGlobalOptionReturnsOptionValueWhenOptionExists() : void
     {
         $option = $this->options->getGlobalOption('test');