+++ CHANGELOG.md
@@ -3,6 +3,19 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 6.0.0
+
+### Changed
+
+- `PMG\Queue\Driver\PheanstalkDriver` no longer accepts arrays as its `$options`
+  argument, instead pass `PMG\Queue\Driver\Pheanstalk\ArrayOptions`.
+- `PMG\Queue\Driver\Pheanstalk\BuryFailureStrategy` no longer accepts an integer
+  argument to its constructor, pass an options instance instead.
+- `PMG\Queue\Driver\Pheanstalk\PheanstalkOptions` now has `int|string|bool`
+  return types for its `get{Global,Message}Option` methods.
+- `PMG\Queue\Driver\Pheanstalk\FailureStrategy` now has `void` return type
+  on its `fail` method.
+
 ## 5.2.0
 
 ### Added